modify home page, to acess information from App.jsx
